really?

+++ source/client_channel_handler.c
@@ -85,7 +85,7 @@ static int s_packet_handler_connack(
     { /* BEGIN CRITICAL SECTION */
         mqtt_connection_lock_synced_data(connection);
         /* User requested disconnect, don't do anything */
-        if (connection->synced_data.state >= AWS_MQTT_CLIENT_STATE_DISCONNECTING) {
+        if ((int)connection->synced_data.state >= (int)AWS_MQTT_CLIENT_STATE_DISCONNECTING) {
             mqtt_connection_unlock_synced_data(connection);
             AWS_LOGF_TRACE(
                 AWS_LS_MQTT_CLIENT, "id=%p: User has requested disconnect, dropping connection", (void *)connection);
